requirements.txt
ffmpeg-python
jsonschema
PyYAML
plexapi